pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pet-infra"
version = "2.5.0"
description = "Shared infrastructure for Train-Pet-Pipeline"
requires-python = ">=3.11,<3.12"
dependencies = [
    "tenacity>=8.0,<9.0",
    "pyyaml>=6.0,<7.0",
    "pet-schema @ git+https://github.com/Train-Pet-Pipeline/pet-schema.git@v2.4.0",
    "mmengine-lite>=0.10,<0.11",
    "hydra-core>=1.3,<1.4",
    "hydra-zen>=0.15,<0.17",
    "click>=8.1,<9.0",
    "networkx>=3.2,<4.0",
    "clearml>=1.14,<2.0",
    "boto3>=1.34",
    "requests>=2.31",
]

[project.optional-dependencies]
gpu = ["torch>=2.1"]
api = ["httpx>=0.27,<1.0"]
dev = [
    "pytest>=8.0,<9.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "types-requests>=2.31",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "moto[s3]>=5.0",
]
sync = [
    "tomli>=2.0; python_version < '3.12'",
    "tomli_w>=1.0",
]

[project.scripts]
pet = "pet_infra.cli:main"

[project.entry-points."pet_infra.plugins"]
pet_infra = "pet_infra._register:register_all"

[project.entry-points."pet_infra.experiment_loggers"]
clearml = "pet_infra.experiment_logger.clearml_logger:ClearMLLogger"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
